Move preprocess to package folder
